+++ packages/vm/src/errors.rs
@@ -55,13 +55,13 @@ pub enum VmError {
     },
     #[snafu(display("Region length too big. Got {}, limit {}", length, max_length))]
     // Note: this only checks length, not capacity
-    RegionLengthTooBigErr {
+    RegionLengthTooBig {
         length: usize,
         max_length: usize,
         backtrace: snafu::Backtrace,
     },
     #[snafu(display("Region too small. Got {}, required {}", size, required))]
-    RegionTooSmallErr {
+    RegionTooSmall {
         size: usize,
         required: usize,
         backtrace: snafu::Backtrace,
@@ -109,6 +109,14 @@ pub fn make_integrity_err() -> VmError {
     IntegrityErr {}.build()
 }
 
+pub fn make_region_length_too_big(length: usize, max_length: usize) -> VmError {
+    RegionLengthTooBig { length, max_length }.build()
+}
+
+pub fn make_region_too_small(size: usize, required: usize) -> VmError {
+    RegionTooSmall { size, required }.build()
+}
+
 pub fn make_runtime_err<S: Into<String>>(msg: S) -> VmError {
     RuntimeErr { msg: msg.into() }.build()
 }
@@ -126,6 +134,32 @@ mod test {
         let err = make_cache_err("something went wrong");
         match err {
             VmError::CacheErr { msg, .. } => assert_eq!(msg, "something went wrong"),
+            _ => panic!("Unexpected error"),
+        }
+    }
+
+    #[test]
+    fn make_region_length_too_big_works() {
+        let err = make_region_length_too_big(50, 20);
+        match err {
+            VmError::RegionLengthTooBig {
+                length, max_length, ..
+            } => {
+                assert_eq!(length, 50);
+                assert_eq!(max_length, 20);
+            }
+            _ => panic!("Unexpected error"),
+        }
+    }
+
+    #[test]
+    fn make_region_too_small_works() {
+        let err = make_region_too_small(12, 33);
+        match err {
+            VmError::RegionTooSmall { size, required, .. } => {
+                assert_eq!(size, 12);
+                assert_eq!(required, 33);
+            }
             _ => panic!("Unexpected error"),
         }
     }